Fix Achievement FE calculation

+++ client/js/app.js
@@ -512,9 +512,8 @@ define(["jquery", "storage"], function ($, Storage) {
       this.displayUnlockedAchievement(id);
 
       var nb = parseInt($("#unlocked-achievements").text());
-      const totalNano = parseInt($("#unlocked-nano-achievements").text()) * 100000;
+      const totalNano = parseFloat($("#unlocked-nano-achievements").text()) * 100000;
       $("#unlocked-achievements").text(nb + 1);
-      // @TODO fix bug in calculation
       $("#unlocked-nano-achievements").text((totalNano + nano) / 100000);
     },
